partially #4 ,#6,#7
I have updated the names to what it looks like they should be. I had
some trouble understanding exactly what the names were supposed to be. I
also changed it so I think it will recognize which piece it is. I had
tested the pawn before but I just changed the one test each time instead
of copying it again. All of these changes are untested because I am
using the school computers and the command line uses a different command
to test it that I could not figure out. I will test them later.