Trying to do check
I added a few methods that can take a move that has been approved by
is_valid_move and then change the board to make that move. They then
check if the move puts the king in check. If it does, then it moves the
piece back. If it does not, then it leaves the board updated and
changes turn to the other color.