Finished todo list except for one
I have consolidated most of the code like we did the other day and
added the valid_moves() for the king and pawn and tested them. The pawn
is probably in the least efficient way possible partly because I forgot
the syntax for something and did it the long way. I was not are how to
move directions into their own file.