Simplifying in_check and some slight changes to the pygame chess board.
I simplified the in_check code and it all works. I the last comment you
said to chage the _code_at_square() call to a color_at_square() call in
the king_location method. I did not quit see how to do this and still
know whether the piece was a king or not. I will also be adding my
chess piece images to the repository after I have finished typing this.