2048
I registered for pyohio. OK I got rid of the Tile object. The visual
part now shows what the board object is. I have not tried the animation
yet. It will need to be modified a little to account for double digit
numbers. I added separate methods for each direction of movement. This
may be redundant but I think it will help for now.