motion
ok I got all of the moves to work correctly as far as I can tell except
for one small exception. All of the directions are in their own method
at the moment which is a bit redundant but I was having trouble
combining them. The board displays correctly but I have to change it
show that it can show double digit numbers. I will work on animation
next. The one problem with the tests is that it adds a random piece
after moving.